+++ src/main/java/com/example/repository/AutomobileRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface AutomobileRepository extends JpaRepository<Automobile, Long> {
-    List<Automobile> findByPrezzoLessThan(Double prezzo);
+    List<Automobile> findByPrezzoLessThan(Integer prezzo);
     List<Automobile> findByMarca(String marca);
 }
