+++ src/main/java/com/example/repository/MotociclettaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface MotociclettaRepository extends JpaRepository<Motocicletta, Long> {
-    List<Motocicletta> findByPrezzoLessThan(Double prezzo);
+    List<Motocicletta> findByPrezzoLessThan(Integer prezzo);
     List<Motocicletta> findByMarca(String marca);
 }
